Add semi-functional stub for completion extension
This change includes the beginnings of a completion extension.
The code is quite messy at this point but provides a starting point
for proper code completion.

At present, only a subset of Cobra keywords are provided and all
namespaces for the types contained in referenced assemblies (following
a 'use' statement).

Remaining work include finding a node from a given region and then
including the completion data from that node (e.g. class members).