Use location of source file as working directory when no project exists
When a source file is opened directly as opposed to opening a project,
the working directory is now set to the location of the source file.

bin/Debug folders will be created at this location and will hold the
generated assembly/exe file.  This matches the behavior when openining
an individual C# source file in MonoDevelop.  This also means the user
must have write access to that location.

This resolves issue #15