Ignore parse requests for files that aren't open
This greatly improves project load time.  If we were saving type system
information from the ASTs, then it would make more sense to parse every
file when MonoDevelop starts.  But, since we don't, we only return a
document with folding regions when asked to fully parse a file that is
not yet open.