Add custom completion data list that doesn't autoselect on whitespace
The default behavior for MonoDevelop of selecting the highlighted
entry when entering whitespace doesn't fit well with declaring variables
in Cobra.  This change introduces a new completion data list class
that essentially ignores the user's setting.  Not sure if it's the
"right" thing to do but it's definitely the less annoying thing to do.