Only provide completion for '.', '_', and letters
This better matches user expectation as mentioned in issue #25

Typing...

    i = 5_

should no longer generate completion proposals.

Note that something like...

    x1y2 = "something"
    x_

will still keep the completion window open when the developer hits
the '1' key.  This is desirable behavior.

Unfortunately, the same is not true for Generic types.

    xs = List_

The completion window will disappear upon hitting '<' and then
reappear with odd results when the user starts typing 'of'.  This is
something to be fixed if possible.