Make use of recent enhancements to Cobra compiler library
The CompilerManager now makes use of existing phases when parsing code
and also directly returns a ParsedCobraDocument instance instead of a
list of dynamic objects which must then be inspected by the Parser.

Additionally, AST visitation is now performed immediately after a
CobraModule is added to the newly parsed document.  This eliminates the
possibility of race conditions between the Parser and CompilerManager
that existed with the previous approach.