Partial fix for issue #60
This is the first attempt at parsing generics in support of parameter
completion and mouseover tooltips.  For now, what works is displaying
the correct initializer overloads and their parameters.  However, the
generic parameter name is displayed instead of the specified type.

For example, "foo = List<of String>(" will show the correct initializer
overloads but the tooltip will display "List<of T>" instead of
"List<of String>".

Mouseover tooltips for generic types still aren't working.  We might
need a proper tokenizer and parser instead of the current hackjob.