Workaround for issue #65
Provides the option to disable document/ast caching for large projects.

Set the environment variable MD_COBRA_BIG_PROJECTS to the path to a
.cobraproj file to reduce memory consumed when editing files in that
project.  You can have multiple paths in this environment variable.
On Windows, separate paths with semicolons.  On *nix use colons instead.

This is not a solution to our memory woes, but does help in the meantime
until a proper fix is developed...which could be a while...