Fix bug with completion data list for interfaces
Implicitly public members of an interface will now show up correctly in
the list of members in the completion list for local variables and
fields that are explicitly typed as an interface.

The reason they weren't showing up before is that the .isNames
collection for members coming from an interface was empty instead of
having an explicit entry for "public".