Handle MonoDevelop.Projects.Formats.MSBuild becoming library
The MonoDevelop.Projects.Formats.MSBuild was changed from an x86
executable assembly to an anycpu library.  This means we no longer need
to install Cobra targetting x86 on Windows 64-bit anymore to
successfully compile the addin.  Since the assembly is now compiled for
anycpu, all references can loaded using reflection in either a 32-bit or
64-bit process.